+++ @cybernaut/chrome/types/is-png.d.ts
@@ -1,5 +1,0 @@
-declare module 'is-png' {
-  function isPNG(data: Buffer): boolean;
-
-  export = isPNG;
-}