+++ @cybernaut/core/types/deep-strict-equal.d.ts
@@ -1,6 +1,0 @@
-declare module 'deep-strict-equal' {
-  /* tslint:disable-next-line no-any */
-  function deepStrictEqual(a: any, b: any): boolean;
-
-  export = deepStrictEqual;
-}