+++ @cybernaut/test/src/sleep.ts
@@ -1,0 +1,3 @@
+export async function sleep(duration: number): Promise<void> {
+  return new Promise<void>(resolve => setTimeout(resolve, duration));
+}