+++ @cybernaut/types/src/Action.ts
@@ -1,6 +1,0 @@
-import {Implementation} from './Implementation';
-
-export interface Action<T> {
-  readonly description: string;
-  readonly implementation: Implementation<T>;
-}