+++ @cybernaut/types/src/Implementation.ts
@@ -1,1 +1,0 @@
-export type Implementation<T> = () => Promise<T>;